Switching from Linux to Windows

+++ admin/src/redux/actions/user.js
@@ -4,6 +4,7 @@ import {
     end,
     registerReducer,
     loginReducer,
+    logoutReducer,
     getUsersReducer,
     getUserReducer,
     getUserStatsReducer,
@@ -13,81 +14,98 @@ import {
 import * as api from '../api/index'
 import Cookie from 'js-cookie'
 
-export const register = (userData, navigate) => async (dispatch) => {
-    dispatch(start())
+export const register = (userData, navigate, toUser) => async (dispatch) => {
     try {
+        dispatch(start())
         await api.register(userData)
         dispatch(registerReducer())
-        navigate('/auth/login')
+        toUser ? navigate('/auth/login') : navigate('/users')
+        dispatch(end())
     } catch (err) {
+        console.log(err)
         dispatch(error(err.message))
     }
-    dispatch(end())
 }
 
 export const login = (userData, navigate) => async (dispatch) => {
-    dispatch(start())
     try {
+        dispatch(start())
         const { data } = await api.login(userData)
         dispatch(loginReducer(data.result))
-        Cookie.set('profile', JSON.stringify(action.payload))
         navigate('/')
+        Cookie.set('profile', JSON.stringify(data.result))
+        dispatch(end())
     } catch (err) {
         dispatch(error(err.message))
     }
-    dispatch(end())
+}
+
+export const logout = (navigate) => async (dispatch) => {
+    try {
+        dispatch(start())
+        Cookie.remove('profile')
+        dispatch(logoutReducer())
+        navigate('/auth/login')
+        dispatch(end())
+    } catch (err) {
+        dispatch(error(err.message))
+    }
 }
 
 
 
 
 export const getUsers = () => async (dispatch) => {
-    dispatch(start())
     try {
+        dispatch(start())
         const { data } = await api.getUsers()
+        console.log('data', data)
         dispatch(getUsersReducer(data.result))
+        dispatch(end())
     } catch (err) {
-         dispatch(error(err.message))
+        dispatch(error(err.message))
     }
-    dispatch(end())
 }
 export const getUser = (userId) => async (dispatch) => {
-    dispatch(start())
     try {
+        dispatch(start())
         const { data } = await api.getUser(userId)
         dispatch(getUserReducer(data.result))
+        dispatch(end())
     } catch (err) {
         dispatch(error(err.message))
     }
-    dispatch(end())
 }
 export const getUserStats = () => async (dispatch) => {
-    dispatch(start())
     try {
+        dispatch(start())
         const { data } = await api.getUserStats()
         dispatch(getUserStatsReducer(data.result))
+        dispatch(end())
     } catch (err) {
         dispatch(error(err.message))
     }
-    dispatch(end())
 }
-export const updateUser = () => async (dispatch) => {
-    dispatch(start())
+export const updateUser = (userId, userData, navigate) => async (dispatch) => {
     try {
-        const { data } = await api.updateUser()
+        dispatch(start())
+        const { data } = await api.updateUser(userId, userData)
         dispatch(updateUserReducer(data.result))
+        navigate('/users')
+        dispatch(end())
     } catch (err) {
+        console.log(err)
         dispatch(error(err.message))
     }
-    dispatch(end())
 }
-export const deleteUser = () => async (dispatch) => {
-    dispatch(start())
+export const deleteUser = (userId, navigate) => async (dispatch) => {
     try {
-        const { data } = await api.deleteUser()
+        dispatch(start())
+        const { data } = await api.deleteUser(userId)
         dispatch(deleteUserReducer(data.result))
+        navigate('/users')
+        dispatch(end())
     } catch (err) {
         dispatch(error(err.message))
     }
-    dispatch(end())
 }
